datenstruktur landscape 2d gemacht

+++ src/Field.java
@@ -2,9 +2,9 @@ public class Field {
 
     public final int x;
     public final int y;
-    public final double value;
+    public final int value;
 
-    public Field(int x, int y, double value) {
+    public Field(int x, int y, int value) {
         this.x = x;
         this.y = y;
         this.value = value;
